Fix error dialog displaying %2 for device size
In GitLab issue #86 user reported the following error dialog preventing
an operation being composed:
    Could not add this operation to the list
    GParted Bug: A partition cannot end (46374912) after the end of
    the device (%2)
    [                             OK                              ]

Additionally on the command line this warning is reported:
    (gpartedbin:19151): glibmm-WARNING **: 07:48:56.108: invalid
    substitution "%2" in fmt string "A partition cannot end (%1) after
    the end of the device (%2)"

Minimum steps to reproduce are:
    # truncate -s 128035676160 /tmp/disk.img
    # losetup -f --show /tmp/disk.img
    /dev/loop0
    # sfdisk -u S /dev/loop0 << EOF
    2048 39061504 83
    39065598 211003394 5
    0 0 0
    0 0 0
    46376960 8310784 82
    54689792 195379200 83
    EOF
    # partprobe /dev/loop0

Run GParted and attempt to move the start of the extended partition to
the right as much as possible (or any amount of 2 MiB or more).  The
above dialog is displayed.

This change just fixes the error message to display the correct values.
It does not try to fix the cause of the error.  The fixed message
becomes:
    GParted Bug: A partition cannot end (250071039) after the end of
    the device (250069679)